requirements.txt
nltk
pyyaml
wordfreq
